process/requirements.txt
boto3
psycopg2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
botocore
